pyproject.toml
[tool.poetry]
name = "shortdiary-backend"
version = "0.1.0"
description = ""
authors = ["Lukas Martini <hello@lutoma.org>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.1"}
python-multipart = "^0.0.5"
fastapi-jwt-auth = "^0.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
stripe = "^3.5.0"
aerich = "^0.6.3"
confz = "^1.6.1"

[tool.aerich]
tortoise_orm = "main.TORTOISE_ORM"
location = "./src/migrations"
src_folder = "./src"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
